backend/pyproject.toml
[project]
name = "insec-platform-backend"
version = "0.1.0"
description = "insec.ml training lab deployment manager"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy>=2.0",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "email-validator>=2",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "alembic>=1.13",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.32",
    "ruff>=0.6",
    "mypy>=1.11",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "alembic*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "RUF"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
